BACKEND/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.7.1 
ngrok==4.3.1
